don't create `dsp` dir; `dsp` is created by clone
Hi @laumontemayor and @reshama! I appreciate all the work on the pre-work here! Cool to see you extending it! It's diverged considerably from what I would do in terms of organization and style, but don't worry, I'm not going to go through and take off all the trailing whitespace. :smile: Make sure that @jonhanke and @julialintern stay connected to what's happening in `dsp`, since they'll be working with the students produced by it!